Refactor footer component
- Move footer component and styling to "layouts" folder
