+++ src/layouts/Footer.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from './Footer.module.scss';
+import styles from './Layouts.module.scss';
 
 export default function Footer() {
